Make global modal, force user profile edit
