+++ services/manage/app/router.js
@@ -12,7 +12,6 @@ Router.map(function() {
   this.route('logout');
   this.route('authenticate', { path: '/authenticate/:token' });
   this.route('manage', { path: '' }, function() {
-    this.route('profile');
     this.route('orgs', function() {
       this.route('invites');
       this.route('list', { path: '/' }, function() {
