+++ services/manage/app/templates/components/user/modal-form.hbs
@@ -1,16 +1,18 @@
-<BsModal @show={{true}} @backdrop="static" @size="large" @onClose={{action "returnToList"}} as |modal|>
+<BsModal @show={{show}} @backdrop="static" @size="large" @onClose={{action "close" modal.actions.hide}} as |modal|>
 
-  <form {{action this.onSubmit modal.actions.hide on="submit"}}>
+  <form {{action "update" modal.actions.hide on="submit"}}>
     {{#modal.header}}
       <h2 class="modal-title">{{this.title}}</h2>
-      <button
-        type="button"
-        class="close"
-        aria-label="Close"
-        {{action modal.actions.hide}}
-      >
-        {{entypo-icon "cross"}}
-      </button>
+      {{#if isModalCloseable}}
+        <button
+          type="button"
+          class="close"
+          aria-label="Close"
+          {{action modal.actions.hide}}
+        >
+          {{entypo-icon "cross"}}
+        </button>
+      {{/if}}
     {{/modal.header}}
 
     {{#modal.body}}
